Added routes & controllers in backend & integrated in frontend

+++ server/controller/adminController.js
@@ -1,5 +1,6 @@
-import {User} from "../models/authModel.js";
-import {College} from "../models/collegeModel.js";
+import { User } from "../models/authModel.js";
+import { College } from "../models/collegeModel.js";
+import { Review } from "../models/reviewModel.js";
 
 const getAllUsers = async (req, res) => {
   try {
@@ -12,8 +13,14 @@ const getAllUsers = async (req, res) => {
 
 const deleteUser = async (req, res) => {
   try {
-    await User.findByIdAndDelete(req.params.id);
-    res.status(200).json({ message: "User deleted successfully" });
+    const user = await User.findById(req.params.id);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    await user.deleteOne();
+    res
+      .status(200)
+      .json({ message: "User and associated reviews deleted successfully" });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -21,17 +28,15 @@ const deleteUser = async (req, res) => {
 
 const addCollege = async (req, res, next) => {
   try {
-    console.log("name :",req.file.filename)
-  const logo = `/uploads/${req.file.filename}`;
+    const logo = `/uploads/${req.file.filename}`;
     const { name, description, website } = req.body;
-    const collegeExist = await College.findOne({ name: name });
-    
+    const collegeExist = await College.findOne({ name });
+
     if (collegeExist) {
       return res.status(400).json({ msg: "College already exists" });
     }
     const college = await College.create({ name, logo, description, website });
     res.status(201).json(college);
-    console.log("clg is ",college);
   } catch (error) {
     next(error);
   }
@@ -52,18 +57,19 @@ const getCollegeById = async (req, res) => {
     if (college) {
       res.json(college);
     } else {
-      res.status(404).json({ message: 'College not found' });
+      res.status(404).json({ message: "College not found" });
     }
   } catch (error) {
-    res.status(500).json({ message: 'Server error' });
+    res.status(500).json({ message: "Server error" });
   }
 };
-
 
 const updateCollege = async (req, res) => {
   try {
     const { id } = req.params;
-    const { name, logo, description, website } = req.body;
+    const logo = `/uploads/${req.file.filename}`;
+    const { name, description, website } = req.body;
+    console.log("HELLO BACHCHOOO");
 
     const updatedCollege = await College.findByIdAndUpdate(
       id,
@@ -79,12 +85,28 @@ const updateCollege = async (req, res) => {
 
 const deleteCollege = async (req, res) => {
   try {
+    const college = await College.findById(req.params.id);
+    if (!college) {
+      return res.status(404).json({ message: "College not found" });
+    }
+    await Review.deleteMany({ _id: { $in: college.reviews } });
+
     await College.findByIdAndDelete(req.params.id);
-    res.status(200).json({ message: "College deleted successfully" });
+
+    res
+      .status(200)
+      .json({ message: "College and associated reviews deleted successfully" });
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
 };
 
-
-export default {getAllUsers, addCollege, getAllColleges , getCollegeById, updateCollege, deleteCollege, deleteUser}
+export default {
+  getAllUsers,
+  addCollege,
+  getAllColleges,
+  getCollegeById,
+  updateCollege,
+  deleteCollege,
+  deleteUser,
+};
